+++ src/model/transformation/ResizeT.java
@@ -49,7 +49,7 @@ public class ResizeT extends AbstractTransformation {
     if (frame < 0) {
       throw new IllegalArgumentException("Frame cannot be negative.");
     }
-    int newValue = this.getValueAtFrame(frame, this.initialValue, this.finalValue);
+    float newValue = this.getValueAtFrame(frame, this.initialValue, this.finalValue);
     if (this.baseOrHeight == Dimension.BASE) {
       this.shape.resize(newValue, this.shape.getHeight());
     } else {
